+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v3.1.0 (not released)
 
 - Fixed: After item in Content Map is clicked, screen reader not changed to 'Browse mode' and not focus to the corresponding response element. Increase success rate to force SR to switch to 'Browse mode' by set `scrollIntoView.behaviour=instant' and increase wait time before change element focus.
+- Also tested with Brave browser.
 
 ### On Progress